In ASDF system definition for diophantine_system, make name of system the same as the package name.
This commit changes hyphens in the system name to underscores. Using underscores is frowned up
by ASDF (see Chapter 6, "Defining systems with defsystem", in the current manual)
and hyphen is preferred, but it appears that Quicklisp needs for the system name
to be the same as the package name, and I would rather not change the package name.