New script d00521032-with-For_loop_variables.mac to work with values supplied by For_loop_variables.lisp.
For_loop_variables.lisp is not included in this commit because it
is too large (about 80 M). The author posted this link to it:
https://drive.google.com/file/d/17fjlTP0FVPuv9VWXDxXgDeWfnB2Qid5d/view?usp=sharing

Incidentally output some progress messages in distribute_over_tranches.lisp.

d00521032-with-For_loop_variables.mac saves the results (via 'save')
to a file named d00521032-with-For_loop_variables-session-yyyy-mm-ddThh-mm-ssZ.lisp
where yyyy-mm-ddThh-mm-ssZ is a timestamp for the current time in UTC.