New package timedate_middleendian for parsing "middle-endian" date format (e.g. m/d/yyyy).
Some heuristics have been implemented for a "do what I mean" approach.
Overall the parser is lenient about what it accepts; it will try to
make some sense out of the input, in hope of covering typical uses
of middle-endian dates, which are produced by various software packages
in a nonstandardized way.